+++ packages/js-toolkit/utils/random.ts
@@ -1,14 +1,24 @@
 import { isArray, isString } from './is.js';
 
 /**
+ * Get a random number between given bounds.
+ * @param   {number} a   First bound.
+ * @param   {number} [b] Second bound, defaults to 0.
+ * @returns {number} A number between `a` and `b`.
+ */
+export function random(a: number, b = 0): number {
+  return Math.random() * (b - a) + a;
+}
+
+/**
  * Get a random integer between bounds
  *
- * @param {number} a Lower bound
- * @param {number} b Upper bound
- * @returns {number}
+ * @param {number} a First bound.
+ * @param {number} b Second bound.
+ * @returns {number} An integer between `a` and `b`;
  */
 export function randomInt(a: number, b = 0): number {
-  return Math.floor(Math.random() * (a - b + 1)) + b;
+  return Math.round(random(a, b));
 }
 
 /**
@@ -16,6 +26,7 @@ export function randomInt(a: number, b = 0): number {
  *
  * @param {T[] | string} items Array or string
  * @returns {T | undefined}
+ * @throws {Error} Throws an error if `items` is not an array or a string.
  */
 export function randomItem<T>(items: T[] | string): T | string | undefined {
   if (!isArray(items) && !isString(items)) {